fixed dot replacement for more than 2 dots in prefix added mocks and tests
