+++ package.json
@@ -8,7 +8,7 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "mocha --recursive './tests/*test.js'",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
     "build": "./build.sh"
   },
   "repository": {
@@ -28,7 +28,13 @@
   "homepage": "https://github.com/Jabb0/CatchAllBird#readme",
   "devDependencies": {
     "chai": "^4.3.7",
+    "jest": "^29.7.0",
     "mocha": "^10.2.0",
     "uglify-js": "^3.17.4"
+  },
+  "jest": {
+    "setupFiles": ["./jest.setup.js"],
+    "testEnvironment": "node",
+    "transform": {}
   }
 }
